// src/api/generated/model/todoResponse.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * toodl-api
 * Toodl Api
 * OpenAPI spec version: 2.0.0
 */

export interface TodoResponse {
  id: number;
  subject: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  enableDeadline: boolean | null;
  /** @nullable */
  isAllDay: boolean | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  recurrenceRule: string | null;
  /** @nullable */
  startTimezone: string | null;
  /** @nullable */
  endTimezone: string | null;
  /** @nullable */
  startTime: string | null;
  /** @nullable */
  endTime: string | null;
  /** @nullable */
  recurrenceException: string | null;
  position: string;
  done: boolean;
  /** @nullable */
  listId: number | null;
  userId: number;
  /** @nullable */
  categoryId: number | null;
  createdAt: string;
  updatedAt: string;
}